Allow to pass http basic credentials during configuration
This is very useful when building the sites on dev machine or using an
external maps service.

Note that the latter use case is still not completely supported by qwc2,
because some runtime requests like getFeatureInfo needs proper CORS
configuration if the front has a different Origin than the map server.