Update renderLayers in Map.jsx
In the renderLayer function the zIndex is set in the loop starting from 0, incrementally increased and set as options.zIndex. 
The layer.zIndex can be defined in MeasurementSupport.jsx or EditingSupport.jsx and others, but is not used to render the layers. So the layer order isn't influenced by setting layer.zIndex. Due to that, the marker-icon from height profile renders underneath the profile line.