Search provider text update in Searchbox
Actually, when I click on a recent search it updates the text in the Searchbox. 
This PR aims to implement the same logic for the searchproviders results. 
It would be a great addition for the UX to do so.

I made a new function "updatetext" because I do not want the layer to get unselected if I click again, but I still want it to get unselected If I remove text.

I Don't know if my logic for this is good, do not hesitate to inform me If I need to change anything. 

Regards, Clément.