Update Print.jsx
Prevents certain print layouts from being displayed in the layout list of the print dialog.

Several print layouts are defined for a QWC theme. Some of them shouldn't be displayed in the layout list of the print dialog, but are used to print an info portfolio site (special reports created with a custom service).
Therefore, I can't exclude these print layouts in QGIS.

So I defined a prefix for print layout names. All layouts starting with these prefix are not displayed in the print dialog.
The prefix is ​​specified in the config.json file. I also specified a sort order and a default layout:

Added plugin config variables:
1. hidePrintlayoutPrefix: All layouts with this prefix will not be displayed in the print layoutlist
2. sortPrintlayouts: Alphabetical order for print layouts: asc or desc
3. defaultPrintlayout: The default layout to display first

config.json:
{
  "name": "Print",
  "cfg":  #{
    …
    "hidePrintlayoutPrefix": "_hide",
    "sortPrintlayouts": "desc",
    "defaultPrintlayout": "A4 - Hochformat"
  }
}